Rename service worker registration file
