+++ react-app/src/index.js
@@ -8,7 +8,7 @@ import ScrollToTop from './utils/RouteHelper';
 import {BrowserRouter} from 'react-router-dom';
 import { Redirect, Route } from 'react-router';
 import isPublishInstance from './utils/is-publish-instance';
-import { register } from './utils/service-worker';
+import { register } from './utils/register-service-worker';
 
 function render(model) {
     ReactDOM.render((
