+++ react-app/src/utils/register-service-worker.js
